+++ app/models/pay/customer.rb
@@ -13,8 +13,6 @@ module Pay
     validates :processor, presence: true
     validates :processor_id, allow_blank: true, uniqueness: {scope: :processor, case_sensitive: true}
 
-    attribute :payment_method_token, :string
-
     # Account(s) for marketplace payments
     store_accessor :data, :braintree_account
 
@@ -23,23 +21,13 @@ module Pay
     store_accessor :data, :currency
 
     delegate :email, to: :owner
-    delegate_missing_to :pay_processor
 
-    %w[stripe braintree paddle_billing paddle_classic shopify fake_processor].each do |processor_name|
+    %w[stripe braintree paddle_billing paddle_classic lemon_squeezy shopify fake_processor].each do |processor_name|
       scope processor_name, -> { where(processor: processor_name) }
 
       define_method :"#{processor_name}?" do
         processor == processor_name
       end
-    end
-
-    def self.pay_processor_for(name)
-      "Pay::#{name.to_s.classify}::Billable".constantize
-    end
-
-    def pay_processor
-      return if processor.blank?
-      @pay_processor ||= self.class.pay_processor_for(processor).new(self)
     end
 
     def update_payment_method(payment_method_id)
@@ -51,8 +39,7 @@ module Pay
     end
 
     def subscribed?(name: Pay.default_product_name, processor_plan: nil)
-      query = {name: name, processor_plan: processor_plan}.compact
-      subscriptions.active.where(query).exists?
+      subscriptions.active.where({name: name, processor_plan: processor_plan}.compact).exists?
     end
 
     def on_trial?(name: Pay.default_product_name, plan: nil)