+++ lib/pay.rb
@@ -20,6 +20,7 @@ module Pay
   autoload :Shopify, "pay/shopify"
   autoload :PaddleBilling, "pay/paddle_billing"
   autoload :PaddleClassic, "pay/paddle_classic"
+  autoload :LemonSqueezy, "pay/lemon_squeezy"
   autoload :Stripe, "pay/stripe"
 
   autoload :Webhooks, "pay/webhooks"
@@ -57,7 +58,7 @@ module Pay
   @@routes_path = "/pay"
 
   mattr_accessor :enabled_processors
-  @@enabled_processors = [:stripe, :braintree, :paddle_billing, :paddle_classic, :shopify]
+  @@enabled_processors = [:stripe, :braintree, :paddle_billing, :paddle_classic, :lemon_squeezy, :shopify]
 
   mattr_accessor :send_emails
   @@send_emails = true
@@ -129,6 +130,16 @@ module Pay
       option.call(*remaining_args)
     else
       option
+    end
+  end
+
+  def self.sync(params)
+    if (session_id = params[:stripe_checkout_session_id] || params[:session_id])
+      Pay::Stripe.sync_checkout_session(session_id)
+    elsif (transaction_id = params[:paddle_billing_transaction_id] || params[:transaction_id])
+      Pay::PaddleBilling.sync_transaction(transaction_id)
+    elsif (order_id = params[:lemon_squeezy_order_id])
+      Pay::LemonSqueezy.sync_order(order_id)
     end
   end
 end