+++ lib/pay/lemon_squeezy/webhooks/order.rb
@@ -1,0 +1,11 @@
+module Pay
+  module LemonSqueezy
+    module Webhooks
+      class Order
+        def call(order)
+          Pay::LemonSqueezy.sync_order(order.id, object: order)
+        end
+      end
+    end
+  end
+end