+++ lib/pay/paddle_classic.rb
@@ -1,10 +1,8 @@
 module Pay
   module PaddleClassic
-    autoload :Billable, "pay/paddle_classic/billable"
-    autoload :Charge, "pay/paddle_classic/charge"
-    autoload :Error, "pay/paddle_classic/error"
-    autoload :PaymentMethod, "pay/paddle_classic/payment_method"
-    autoload :Subscription, "pay/paddle_classic/subscription"
+    class Error < Pay::Error
+      delegate :message, to: :cause
+    end
 
     module Webhooks
       autoload :SignatureVerifier, "pay/paddle_classic/webhooks/signature_verifier"
@@ -20,7 +18,8 @@ module Pay
     def self.enabled?
       return false unless Pay.enabled_processors.include?(:paddle_classic) && defined?(::Paddle)
 
-      Pay::Engine.version_matches?(required: "~> 2.1", current: ::Paddle::VERSION) || (raise "[Pay] paddle gem must be version ~> 2.1")
+      Pay::Engine.version_matches?(required: "~> 2.5",
+        current: ::Paddle::VERSION) || (raise "[Pay] paddle gem must be version ~> 2.5")
     end
 
     def self.client
@@ -73,9 +72,12 @@ module Pay
       Pay::Webhooks.configure do |events|
         events.subscribe "paddle_classic.subscription_created", Pay::PaddleClassic::Webhooks::SubscriptionCreated.new
         events.subscribe "paddle_classic.subscription_updated", Pay::PaddleClassic::Webhooks::SubscriptionUpdated.new
-        events.subscribe "paddle_classic.subscription_cancelled", Pay::PaddleClassic::Webhooks::SubscriptionCancelled.new
-        events.subscribe "paddle_classic.subscription_payment_succeeded", Pay::PaddleClassic::Webhooks::SubscriptionPaymentSucceeded.new
-        events.subscribe "paddle_classic.subscription_payment_refunded", Pay::PaddleClassic::Webhooks::SubscriptionPaymentRefunded.new
+        events.subscribe "paddle_classic.subscription_cancelled",
+          Pay::PaddleClassic::Webhooks::SubscriptionCancelled.new
+        events.subscribe "paddle_classic.subscription_payment_succeeded",
+          Pay::PaddleClassic::Webhooks::SubscriptionPaymentSucceeded.new
+        events.subscribe "paddle_classic.subscription_payment_refunded",
+          Pay::PaddleClassic::Webhooks::SubscriptionPaymentRefunded.new
       end
     end
   end