+++ test/pay/braintree/subscription_test.rb
@@ -2,8 +2,8 @@ require "test_helper"
 
 class Pay::Braintree::SubscriptionTest < ActiveSupport::TestCase
   setup do
-    @pay_customer = Pay::Customer.create!(processor: :braintree, owner: users(:none))
-    @pay_customer.payment_method_token = "fake-valid-visa-nonce"
+    @pay_customer = Pay::Braintree::Customer.create!(processor: :braintree, owner: users(:none))
+    @pay_customer.update_payment_method "fake-valid-visa-nonce"
   end
 
   test "braintree cancel" do
@@ -14,7 +14,7 @@ class Pay::Braintree::SubscriptionTest < ActiveSupport::TestCase
       assert_equal "active", @subscription.status
       assert @subscription.on_grace_period?
       assert @subscription.active?
-      assert_equal @subscription.ends_at.to_date, @subscription.processor_subscription.billing_period_end_date.to_date
+      assert_equal @subscription.ends_at.to_date, @subscription.api_record.billing_period_end_date.to_date
     end
   end
 
@@ -48,6 +48,9 @@ class Pay::Braintree::SubscriptionTest < ActiveSupport::TestCase
       pay_subscription = @pay_customer.subscription
       pay_subscription.cancel_now!
 
+      # Account for time hitting the API because we've frozen time and ends_at is ahead
+      travel 2.seconds
+
       # Canceling during a trial ends the subscription, but continues to give access during the trial period
       assert_equal "canceled", pay_subscription.status
       refute pay_subscription.active?
@@ -59,7 +62,7 @@ class Pay::Braintree::SubscriptionTest < ActiveSupport::TestCase
 
   test "braintree processor subscription" do
     @pay_customer.subscribe(trial_period_days: 0)
-    assert_equal @pay_customer.subscription.processor_subscription.class, Braintree::Subscription
+    assert_equal @pay_customer.subscription.api_record.class, Braintree::Subscription
     assert_equal "active", @pay_customer.subscription.status
   end
 
@@ -67,7 +70,7 @@ class Pay::Braintree::SubscriptionTest < ActiveSupport::TestCase
     @pay_customer.subscribe(plan: "default", trial_period_days: 0)
     @pay_customer.subscription.swap("big")
 
-    assert_equal "big", @pay_customer.subscription.processor_subscription.plan_id
+    assert_equal "big", @pay_customer.subscription.api_record.plan_id
     assert_equal "active", @pay_customer.subscription.status
   end
 
@@ -75,7 +78,7 @@ class Pay::Braintree::SubscriptionTest < ActiveSupport::TestCase
     @pay_customer.subscribe(plan: "default", trial_period_days: 0)
     @pay_customer.subscription.swap("yearly")
 
-    assert_equal "yearly", @pay_customer.subscription.processor_subscription.plan_id
+    assert_equal "yearly", @pay_customer.subscription.api_record.plan_id
     assert_equal "active", @pay_customer.subscription.status
   end