+++ test/pay/stripe/subscription_test.rb
@@ -21,20 +21,20 @@ class Pay::Stripe::SubscriptionTest < ActiveSupport::TestCase
     refute pay_subscription.active?
   end
 
-  test "change stripe subscription quantity" do
-    @pay_customer.processor_id = nil
-    @pay_customer.payment_method_token = "pm_card_visa"
+  test "stripe change subscription quantity" do
+    @pay_customer.update(processor_id: nil)
+    @pay_customer.update_payment_method "pm_card_visa"
     subscription = @pay_customer.subscribe(name: "default", plan: "default")
     subscription.change_quantity(5)
-    stripe_subscription = subscription.processor_subscription
+    stripe_subscription = subscription.api_record
     assert_equal 5, stripe_subscription.quantity
     assert_equal 5, subscription.quantity
   end
 
   test "cancel_now when scheduled for cancellation" do
     travel_to(VCR.current_cassette&.originally_recorded_at || Time.current) do
-      @pay_customer.processor_id = nil
-      @pay_customer.payment_method_token = "pm_card_visa"
+      @pay_customer.update(processor_id: nil)
+      @pay_customer.update_payment_method "pm_card_visa"
       subscription = @pay_customer.subscribe(name: "default", plan: "default")
       subscription.cancel
       assert subscription.active?
@@ -47,14 +47,14 @@ class Pay::Stripe::SubscriptionTest < ActiveSupport::TestCase
     end
   end
 
-  test "change stripe subscription quantity with nil subscription items" do
+  test "stripe change subscription quantity with nil subscription items" do
     pay_subscription = Pay::Stripe::Subscription.sync("123", object: fake_stripe_subscription)
     pay_subscription.update!(subscription_items: nil)
     ::Stripe::Subscription.stubs(:update)
     assert pay_subscription.change_quantity(5)
   end
 
-  test "change stripe subscription quantity with [] subscription items" do
+  test "stripe change subscription quantity with [] subscription items" do
     pay_subscription = Pay::Stripe::Subscription.sync("123", object: fake_stripe_subscription)
     pay_subscription.update!(subscription_items: [])
     ::Stripe::Subscription.stubs(:update)
@@ -166,14 +166,23 @@ class Pay::Stripe::SubscriptionTest < ActiveSupport::TestCase
 
   test "stripe resume on grace period" do
     travel_to(VCR.current_cassette&.originally_recorded_at || Time.current) do
-      @pay_customer.processor_id = nil
-      @pay_customer.payment_method_token = "pm_card_visa"
+      @pay_customer.update(processor_id: nil)
+      @pay_customer.update_payment_method "pm_card_visa"
       subscription = @pay_customer.subscribe(name: "default", plan: "default")
       subscription.cancel
       assert_not_nil subscription.ends_at
       subscription.resume
       assert_nil subscription.ends_at
       assert_equal "active", subscription.status
+    end
+  end
+
+  test "it will throw an error if subscription cannot be resumed" do
+    Pay::Stripe::Subscription.sync("123", object: fake_stripe_subscription)
+
+    refute @pay_customer.subscription.resumable?
+    assert_raises Pay::Error do
+      @pay_customer.subscription.resume
     end
   end
 
@@ -315,18 +324,18 @@ class Pay::Stripe::SubscriptionTest < ActiveSupport::TestCase
     assert_equal payment_method, pay_subscription.payment_method
   end
 
-  test "change stripe subscription default payment method" do
-    @pay_customer.processor_id = nil
-    @pay_customer.payment_method_token = "pm_card_visa"
+  test "stripe change subscription default payment method" do
+    @pay_customer.update(processor_id: nil)
+    @pay_customer.update_payment_method "pm_card_visa"
     subscription = @pay_customer.subscribe(name: "default", plan: "default")
 
     payment_method = ::Stripe::PaymentMethod.attach("pm_card_discover", {customer: @pay_customer.processor_id})
     subscription.update_payment_method payment_method.id
     assert_equal payment_method.id, subscription.payment_method_id
-    assert_equal payment_method.id, subscription.processor_subscription.default_payment_method.id
+    assert_equal payment_method.id, subscription.api_record.default_payment_method.id
 
     subscription.update_payment_method ""
     assert_nil subscription.payment_method_id
-    assert_nil subscription.processor_subscription.default_payment_method
+    assert_nil subscription.api_record.default_payment_method
   end
 end